requirements.txt
fastapi==0.111.0
itsdangerous==2.2.0
httpx==0.27.0
redis==5.0.7
fastapi-mqtt==2.2.0
paho-mqtt==2.1.0
